refactor: merge AgentConfigGeneratorUseCase into AgentUseCase
Fold the AI generation usecase into AgentUseCase rather than keeping a
separate class: AgentUseCase now takes an AiGenerator as a third
constructor dependency, and create/update become generateAgentInput/
reviseAgentInput to avoid colliding with the existing createHermesAgent/
updateHermesAgent CRUD methods. Tests and the agent-config router are
updated accordingly; the agentConfig.create/update tRPC route names are
unchanged.
