+++ apps/dashboard/src/server/routers/trpc/agent-config.ts
@@ -1,26 +1,23 @@
 import { z } from "zod";
 
 import { AgentInput, AgentInputObjectSchema } from "@/entities";
-import {
-  AgentConfigGeneratorUseCase,
-  PromptSchema,
-} from "@/server/usecases/agent-config-generator";
+import { AgentUseCase, PromptSchema } from "@/server/usecases/agent";
 
 import { protectedProcedure, t } from "./shared.js";
 
-const usecase = new AgentConfigGeneratorUseCase();
+const usecase = new AgentUseCase();
 
 export const agentConfigRouter = t.router({
   create: protectedProcedure
     .input(z.object({ prompt: PromptSchema }))
     .mutation(async ({ ctx, input }): Promise<AgentInput> => {
-      return await usecase.create(ctx, input.prompt);
+      return await usecase.generateAgentInput(ctx, input.prompt);
     }),
 
   update: protectedProcedure
     // AgentInputObjectSchema (no superRefine) so half-finished drafts are accepted.
     .input(z.object({ prompt: PromptSchema, config: AgentInputObjectSchema }))
     .mutation(async ({ ctx, input }): Promise<AgentInput> => {
-      return await usecase.update(ctx, input.config, input.prompt);
+      return await usecase.reviseAgentInput(ctx, input.config, input.prompt);
     }),
 });
