+++ apps/dashboard/src/server/usecases/agent.ts
@@ -3,9 +3,13 @@ import { z } from "zod";
 
 import {
   Agent,
+  AgentEnvVar,
   AgentInput,
+  AgentInputObjectSchema,
   AgentInputSchema,
   Context,
+  ENV_PLACEHOLDER_SENTINEL,
+  ENV_SECRET_SENTINEL,
   Env,
   JsonPatchOp,
   Skill,
@@ -13,8 +17,10 @@ import {
   WebhookVariables,
 } from "@/entities";
 
+import { AiSdkGenerator } from "../infras/ai-sdk";
 import { KubernetesClient } from "../infras/kubernetes/client";
 import { LocalConfig } from "../infras/local-hermeum-config";
+import { AiGenerator } from "./adaptors/generator";
 import { Runtime } from "./adaptors/runtime";
 import { ConfigAdaptor } from "./adaptors/config";
 
@@ -23,10 +29,30 @@ export const ListAgentsFilterSchema = z.object({
 });
 export type ListAgentsFilter = z.infer<typeof ListAgentsFilterSchema>;
 
+export const PromptSchema = z.string().min(1).max(4000);
+export type Prompt = z.infer<typeof PromptSchema>;
+
+// Field semantics live in the output schema's .describe() texts; this prompt
+// only carries the task framing and cross-field rules.
+export const AGENT_INPUT_SYSTEM_PROMPT = `You generate Hermes agent definitions — a JSON
+object that prefills the form for an autonomous Hermes agent deployed to
+Kubernetes. Field semantics are defined by the output schema.
+
+Rules:
+- Credentials in env must have sensitive: true and the literal value
+  "${ENV_PLACEHOLDER_SENTINEL}" — never invent or guess secret values.
+- When config.platforms.webhook.enabled is true, env MUST include a sensitive
+  WEBHOOK_SECRET; when config.api_server.enabled is true, a sensitive
+  API_SERVER_KEY (both with value "${ENV_PLACEHOLDER_SENTINEL}").
+- Sensitive values shown as "${ENV_SECRET_SENTINEL}" in an existing definition are stored
+  secrets — keep them as the literal "${ENV_SECRET_SENTINEL}" when revising.
+- Only enable features the request calls for. Prefer minimal, valid output.`;
+
 export class AgentUseCase {
   constructor(
     private readonly runtime: Runtime = new KubernetesClient(),
-    private readonly config: ConfigAdaptor = new LocalConfig()
+    private readonly config: ConfigAdaptor = new LocalConfig(),
+    private readonly generator: AiGenerator = new AiSdkGenerator()
   ) {}
 
   getmutatingWebhookJsonPatch(agent: Agent): JsonPatchOp[] | null {
@@ -204,5 +230,70 @@ export class AgentUseCase {
     if (ctx.user!.id !== resource.userId) {
       throw new Error("You don't have permission to perform this action");
     }
+  }
+
+  async generateAgentInput(ctx: Context, prompt: Prompt): Promise<AgentInput> {
+    prompt = PromptSchema.parse(prompt);
+    const output = await this.generator.generateAgentInput({
+      system: AGENT_INPUT_SYSTEM_PROMPT,
+      prompt: `Create a new Hermes agent definition from this request:\n\n${prompt}`,
+    });
+    return this.finalizeGeneratedAgentInput(output);
+  }
+
+  async reviseAgentInput(ctx: Context, current: AgentInput, prompt: Prompt): Promise<AgentInput> {
+    prompt = PromptSchema.parse(prompt);
+    const output = await this.generator.generateAgentInput({
+      system: AGENT_INPUT_SYSTEM_PROMPT,
+      prompt:
+        "Here is an existing Hermes agent definition as JSON:\n\n" +
+        JSON.stringify(current, null, 2) +
+        "\n\nApply the following change and return the FULL revised definition " +
+        "(keep everything not affected by the change unchanged):\n\n" +
+        prompt,
+    });
+    return this.finalizeGeneratedAgentInput(output, current.env ?? []);
+  }
+
+  private finalizeGeneratedAgentInput(
+    output: AgentInput,
+    existingEnv: readonly AgentEnvVar[] = []
+  ): AgentInput {
+    const parsed = AgentInputObjectSchema.parse(output);
+    return { ...parsed, env: this.scrubGeneratedEnv(parsed, existingEnv) };
+  }
+
+  // The system prompt asks for these invariants, but LLMs occasionally slip;
+  // patch the output rather than reject it so the prefill stays usable.
+  private scrubGeneratedEnv(
+    parsed: AgentInput,
+    existingEnv: readonly AgentEnvVar[]
+  ): AgentEnvVar[] | undefined {
+    const existingSensitive = new Set(
+      existingEnv.filter((v) => v.sensitive === true).map((v) => v.name)
+    );
+
+    // Never pass through generated secret values: sensitive vars may only hold
+    // the fill-me placeholder, or "<secret>" when it round-trips an already
+    // stored secret (the kubernetes client reuses the stored value for it).
+    const env = (parsed.env ?? []).map((v) => {
+      if (!v.sensitive || v.value === ENV_PLACEHOLDER_SENTINEL) return v;
+      if (v.value === ENV_SECRET_SENTINEL && existingSensitive.has(v.name)) return v;
+      return { ...v, value: ENV_PLACEHOLDER_SENTINEL };
+    });
+
+    const requireSensitiveEnv = (name: string) => {
+      if (!env.some((v) => v.name === name && v.sensitive === true)) {
+        env.push({ name, value: ENV_PLACEHOLDER_SENTINEL, sensitive: true });
+      }
+    };
+    if (parsed.config?.platforms?.webhook?.enabled === true) {
+      requireSensitiveEnv("WEBHOOK_SECRET");
+    }
+    if (parsed.config?.api_server?.enabled === true) {
+      requireSensitiveEnv("API_SERVER_KEY");
+    }
+
+    return env.length > 0 ? env : parsed.env;
   }
 }
